chore: format contracts using forge formatter (#27)

+++ contracts/ERC1155Creator.sol
@@ -17,16 +17,30 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
 
     mapping(uint256 => uint256) private _totalSupply;
 
-    constructor (string memory _name, string memory _symbol) ERC1155Base(_name, _symbol) {}
+    constructor(string memory _name, string memory _symbol) ERC1155Base(_name, _symbol) {}
 
     /**
      * @dev See {IERC165-supportsInterface}.
      */
-    function supportsInterface(bytes4 interfaceId) public view virtual override(ERC1155Core, ERC1155CreatorCore, AdminControl) returns (bool) {
-        return ERC1155CreatorCore.supportsInterface(interfaceId) || ERC1155Core.supportsInterface(interfaceId) || AdminControl.supportsInterface(interfaceId);
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        virtual
+        override(ERC1155Core, ERC1155CreatorCore, AdminControl)
+        returns (bool)
+    {
+        return ERC1155CreatorCore.supportsInterface(interfaceId) || ERC1155Core.supportsInterface(interfaceId)
+            || AdminControl.supportsInterface(interfaceId);
     }
 
-    function _beforeTokenTransfer(address, address from, address to, uint256[] memory ids, uint256[] memory amounts, bytes memory) internal virtual override {
+    function _beforeTokenTransfer(
+        address,
+        address from,
+        address to,
+        uint256[] memory ids,
+        uint256[] memory amounts,
+        bytes memory
+    ) internal virtual override {
         _approveTransfer(from, to, ids, amounts);
     }
 
@@ -41,11 +55,14 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
     /**
      * @dev See {ICreatorCore-registerExtension}.
      */
-    function registerExtension(address extension, string calldata baseURI, bool baseURIIdentical) external override adminRequired {
+    function registerExtension(address extension, string calldata baseURI, bool baseURIIdentical)
+        external
+        override
+        adminRequired
+    {
         requireNonBlacklist(extension);
         _registerExtension(extension, baseURI, baseURIIdentical);
     }
-
 
     /**
      * @dev See {ICreatorCore-unregisterExtension}.
@@ -99,9 +116,11 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
     function setTokenURIExtension(uint256[] calldata tokenIds, string[] calldata uris) external override {
         requireExtension();
         require(tokenIds.length == uris.length, "Invalid input");
-        for (uint i; i < tokenIds.length;) {
+        for (uint256 i; i < tokenIds.length;) {
             _setTokenURIExtension(tokenIds[i], uris[i]);
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
     }
 
@@ -131,9 +150,11 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
      */
     function setTokenURI(uint256[] calldata tokenIds, string[] calldata uris) external override adminRequired {
         require(tokenIds.length == uris.length, "Invalid input");
-        for (uint i; i < tokenIds.length;) {
+        for (uint256 i; i < tokenIds.length;) {
             _setTokenURI(tokenIds[i], uris[i]);
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
     }
 
@@ -147,19 +168,34 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
     /**
      * @dev See {IERC1155CreatorCore-mintBaseNew}.
      */
-    function mintBaseNew(address[] calldata to, uint256[] calldata amounts, string[] calldata uris) public virtual override nonReentrant adminRequired returns(uint256[] memory) {
+    function mintBaseNew(address[] calldata to, uint256[] calldata amounts, string[] calldata uris)
+        public
+        virtual
+        override
+        nonReentrant
+        adminRequired
+        returns (uint256[] memory)
+    {
         return _mintNew(address(0), to, amounts, uris);
     }
 
     /**
      * @dev See {IERC1155CreatorCore-mintBaseExisting}.
      */
-    function mintBaseExisting(address[] calldata to, uint256[] calldata tokenIds, uint256[] calldata amounts) public virtual override nonReentrant adminRequired {
-        for (uint i; i < tokenIds.length;) {
+    function mintBaseExisting(address[] calldata to, uint256[] calldata tokenIds, uint256[] calldata amounts)
+        public
+        virtual
+        override
+        nonReentrant
+        adminRequired
+    {
+        for (uint256 i; i < tokenIds.length;) {
             uint256 tokenId = tokenIds[i];
             require(tokenId > 0 && tokenId <= _tokenCount, "Invalid token");
             require(_tokenExtension(tokenId) == address(0), "Token created by extension");
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
         _mintExisting(address(0), to, tokenIds, amounts);
     }
@@ -167,7 +203,13 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
     /**
      * @dev See {IERC1155CreatorCore-mintExtensionNew}.
      */
-    function mintExtensionNew(address[] calldata to, uint256[] calldata amounts, string[] calldata uris) public virtual override nonReentrant returns(uint256[] memory tokenIds) {
+    function mintExtensionNew(address[] calldata to, uint256[] calldata amounts, string[] calldata uris)
+        public
+        virtual
+        override
+        nonReentrant
+        returns (uint256[] memory tokenIds)
+    {
         requireExtension();
         return _mintNew(msg.sender, to, amounts, uris);
     }
@@ -175,11 +217,18 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
     /**
      * @dev See {IERC1155CreatorCore-mintExtensionExisting}.
      */
-    function mintExtensionExisting(address[] calldata to, uint256[] calldata tokenIds, uint256[] calldata amounts) public virtual override nonReentrant {
+    function mintExtensionExisting(address[] calldata to, uint256[] calldata tokenIds, uint256[] calldata amounts)
+        public
+        virtual
+        override
+        nonReentrant
+    {
         requireExtension();
-        for (uint i; i < tokenIds.length;) {
+        for (uint256 i; i < tokenIds.length;) {
             require(_tokenExtension(tokenIds[i]) == address(msg.sender), "Token not created by this extension");
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
         _mintExisting(msg.sender, to, tokenIds, amounts);
     }
@@ -187,7 +236,10 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
     /**
      * @dev Mint new tokens
      */
-    function _mintNew(address extension, address[] calldata to, uint256[] calldata amounts, string[] calldata uris) internal returns(uint256[] memory tokenIds) {
+    function _mintNew(address extension, address[] calldata to, uint256[] calldata amounts, string[] calldata uris)
+        internal
+        returns (uint256[] memory tokenIds)
+    {
         if (to.length > 1) {
             // Multiple receiver.  Give every receiver the same new token
             tokenIds = new uint256[](1);
@@ -199,12 +251,14 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
         }
 
         // Assign tokenIds
-        for (uint i; i < tokenIds.length;) {
+        for (uint256 i; i < tokenIds.length;) {
             ++_tokenCount;
             tokenIds[i] = _tokenCount;
             // Track the extension that minted the token
             _tokensExtension[_tokenCount] = extension;
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
 
         if (extension != address(0)) {
@@ -212,66 +266,83 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
         }
 
         if (to.length == 1 && tokenIds.length == 1) {
-           // Single mint
-           _mint(to[0], tokenIds[0], amounts[0], new bytes(0));
+            // Single mint
+            _mint(to[0], tokenIds[0], amounts[0], new bytes(0));
         } else if (to.length > 1) {
             // Multiple receivers.  Receiving the same token
             if (amounts.length == 1) {
                 // Everyone receiving the same amount
-                for (uint i; i < to.length;) {
+                for (uint256 i; i < to.length;) {
                     _mint(to[i], tokenIds[0], amounts[0], new bytes(0));
-                    unchecked { ++i; }
+                    unchecked {
+                        ++i;
+                    }
                 }
             } else {
                 // Everyone receiving different amounts
-                for (uint i; i < to.length;) {
+                for (uint256 i; i < to.length;) {
                     _mint(to[i], tokenIds[0], amounts[i], new bytes(0));
-                    unchecked { ++i; }
+                    unchecked {
+                        ++i;
+                    }
                 }
             }
         } else {
             _mintBatch(to[0], tokenIds, amounts, new bytes(0));
         }
 
-        for (uint i; i < tokenIds.length;) {
+        for (uint256 i; i < tokenIds.length;) {
             if (i < uris.length && bytes(uris[i]).length > 0) {
                 _tokenURIs[tokenIds[i]] = uris[i];
             }
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
     }
 
     /**
      * @dev Mint existing tokens
      */
-    function _mintExisting(address extension, address[] calldata to, uint256[] calldata tokenIds, uint256[] calldata amounts) internal {
+    function _mintExisting(
+        address extension,
+        address[] calldata to,
+        uint256[] calldata tokenIds,
+        uint256[] calldata amounts
+    ) internal {
         if (extension != address(0)) {
             _checkMintPermissions(to, tokenIds, amounts);
         }
 
         if (to.length == 1 && tokenIds.length == 1 && amounts.length == 1) {
-             // Single mint
-            _mint(to[0], tokenIds[0], amounts[0], new bytes(0));            
+            // Single mint
+            _mint(to[0], tokenIds[0], amounts[0], new bytes(0));
         } else if (to.length == 1 && tokenIds.length == amounts.length) {
             // Batch mint to same receiver
             _mintBatch(to[0], tokenIds, amounts, new bytes(0));
         } else if (tokenIds.length == 1 && amounts.length == 1) {
             // Mint of the same token/token amounts to various receivers
-            for (uint i; i < to.length;) {
+            for (uint256 i; i < to.length;) {
                 _mint(to[i], tokenIds[0], amounts[0], new bytes(0));
-                unchecked { ++i; }
+                unchecked {
+                    ++i;
+                }
             }
         } else if (tokenIds.length == 1 && to.length == amounts.length) {
             // Mint of the same token with different amounts to different receivers
-            for (uint i; i < to.length;) {
+            for (uint256 i; i < to.length;) {
                 _mint(to[i], tokenIds[0], amounts[i], new bytes(0));
-                unchecked { ++i; }
+                unchecked {
+                    ++i;
+                }
             }
         } else if (to.length == tokenIds.length && to.length == amounts.length) {
             // Mint of different tokens and different amounts to different receivers
-            for (uint i; i < to.length;) {
+            for (uint256 i; i < to.length;) {
                 _mint(to[i], tokenIds[i], amounts[i], new bytes(0));
-                unchecked { ++i; }
+                unchecked {
+                    ++i;
+                }
             }
         } else {
             revert("Invalid input");
@@ -290,7 +361,12 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
     /**
      * @dev See {IERC1155CreatorCore-burn}.
      */
-    function burn(address account, uint256[] calldata tokenIds, uint256[] calldata amounts) public virtual override nonReentrant {
+    function burn(address account, uint256[] calldata tokenIds, uint256[] calldata amounts)
+        public
+        virtual
+        override
+        nonReentrant
+    {
         require(account == msg.sender || isApprovedForAll(account, msg.sender), "Caller is not owner or approved");
         require(tokenIds.length == amounts.length, "Invalid input");
         if (tokenIds.length == 1) {
@@ -304,35 +380,59 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
     /**
      * @dev See {ICreatorCore-setRoyalties}.
      */
-    function setRoyalties(address payable[] calldata receivers, uint256[] calldata basisPoints) external override adminRequired {
+    function setRoyalties(address payable[] calldata receivers, uint256[] calldata basisPoints)
+        external
+        override
+        adminRequired
+    {
         _setRoyaltiesExtension(address(0), receivers, basisPoints);
     }
 
     /**
      * @dev See {ICreatorCore-setRoyalties}.
      */
-    function setRoyalties(uint256 tokenId, address payable[] calldata receivers, uint256[] calldata basisPoints) external override adminRequired {
+    function setRoyalties(uint256 tokenId, address payable[] calldata receivers, uint256[] calldata basisPoints)
+        external
+        override
+        adminRequired
+    {
         _setRoyalties(tokenId, receivers, basisPoints);
     }
 
     /**
      * @dev See {ICreatorCore-setRoyaltiesExtension}.
      */
-    function setRoyaltiesExtension(address extension, address payable[] calldata receivers, uint256[] calldata basisPoints) external override adminRequired {
+    function setRoyaltiesExtension(
+        address extension,
+        address payable[] calldata receivers,
+        uint256[] calldata basisPoints
+    ) external override adminRequired {
         _setRoyaltiesExtension(extension, receivers, basisPoints);
     }
 
     /**
      * @dev See {ICreatorCore-getRoyalties}.
      */
-    function getRoyalties(uint256 tokenId) external view virtual override returns (address payable[] memory, uint256[] memory) {
+    function getRoyalties(uint256 tokenId)
+        external
+        view
+        virtual
+        override
+        returns (address payable[] memory, uint256[] memory)
+    {
         return _getRoyalties(tokenId);
     }
 
     /**
      * @dev See {ICreatorCore-getFees}.
      */
-    function getFees(uint256 tokenId) external view virtual override returns (address payable[] memory, uint256[] memory) {
+    function getFees(uint256 tokenId)
+        external
+        view
+        virtual
+        override
+        returns (address payable[] memory, uint256[] memory)
+    {
         return _getRoyalties(tokenId);
     }
 
@@ -346,16 +446,16 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
     /**
      * @dev See {ICreatorCore-getFeeBps}.
      */
-    function getFeeBps(uint256 tokenId) external view virtual override returns (uint[] memory) {
+    function getFeeBps(uint256 tokenId) external view virtual override returns (uint256[] memory) {
         return _getRoyaltyBPS(tokenId);
     }
-    
+
     /**
      * @dev See {ICreatorCore-royaltyInfo}.
      */
     function royaltyInfo(uint256 tokenId, uint256 value) external view virtual override returns (address, uint256) {
         return _getRoyaltyInfo(tokenId, value);
-    } 
+    }
 
     /**
      * @dev See {IERC1155MetadataURI-uri}.
@@ -363,7 +463,7 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
     function uri(uint256 tokenId) public view virtual override returns (string memory) {
         return _tokenURI(tokenId);
     }
-    
+
     /**
      * @dev Total amount of tokens in with a given id.
      */
@@ -382,11 +482,17 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
     /**
      * @dev See {ERC1155-_mintBatch}.
      */
-    function _mintBatch(address to, uint256[] memory ids, uint256[] memory amounts, bytes memory data) internal virtual override {
+    function _mintBatch(address to, uint256[] memory ids, uint256[] memory amounts, bytes memory data)
+        internal
+        virtual
+        override
+    {
         super._mintBatch(to, ids, amounts, data);
-        for (uint i; i < ids.length;) {
+        for (uint256 i; i < ids.length;) {
             _totalSupply[ids[i]] += amounts[i];
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
     }
 
@@ -403,9 +509,11 @@ contract ERC1155Creator is AdminControl, ERC1155Base, ERC1155CreatorCore {
      */
     function _burnBatch(address account, uint256[] memory ids, uint256[] memory amounts) internal virtual override {
         super._burnBatch(account, ids, amounts);
-        for (uint i; i < ids.length;) {
+        for (uint256 i; i < ids.length;) {
             _totalSupply[ids[i]] -= amounts[i];
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
     }
