+++ contracts/core/ICreatorCore.sol
@@ -10,7 +10,6 @@ import "openzeppelin/utils/introspection/IERC165.sol";
  * @dev Core creator interface
  */
 interface ICreatorCore is IERC165 {
-
     event ExtensionRegistered(address indexed extension, address indexed sender);
     event ExtensionUnregistered(address indexed extension, address indexed sender);
     event ExtensionBlacklisted(address indexed extension, address indexed sender);
@@ -129,28 +128,33 @@ interface ICreatorCore is IERC165 {
     /**
      * @dev Set royalties of a token
      */
-    function setRoyalties(uint256 tokenId, address payable[] calldata receivers, uint256[] calldata basisPoints) external;
+    function setRoyalties(uint256 tokenId, address payable[] calldata receivers, uint256[] calldata basisPoints)
+        external;
 
     /**
      * @dev Set royalties of an extension
      */
-    function setRoyaltiesExtension(address extension, address payable[] calldata receivers, uint256[] calldata basisPoints) external;
+    function setRoyaltiesExtension(
+        address extension,
+        address payable[] calldata receivers,
+        uint256[] calldata basisPoints
+    ) external;
 
     /**
      * @dev Get royalites of a token.  Returns list of receivers and basisPoints
      */
     function getRoyalties(uint256 tokenId) external view returns (address payable[] memory, uint256[] memory);
-    
+
     // Royalty support for various other standards
     function getFeeRecipients(uint256 tokenId) external view returns (address payable[] memory);
-    function getFeeBps(uint256 tokenId) external view returns (uint[] memory);
+    function getFeeBps(uint256 tokenId) external view returns (uint256[] memory);
     function getFees(uint256 tokenId) external view returns (address payable[] memory, uint256[] memory);
     function royaltyInfo(uint256 tokenId, uint256 value) external view returns (address, uint256);
 
     /**
      * @dev Set the default approve transfer contract location.
      */
-    function setApproveTransfer(address extension) external; 
+    function setApproveTransfer(address extension) external;
 
     /**
      * @dev Get the default approve transfer contract location.
